+++ src/App.tsx
@@ -1,6 +1,6 @@
 import { Calculator } from "./components/Calculator";
-import { ThemeProvider } from "./components/theme-provider";
-import { ThemeToggle } from "./components/theme-toggle";
+import { ThemeProvider } from "./components/theme/theme-provider";
+import { ThemeToggle } from "./components/theme/theme-toggle";
 import "./styles.css";
 
 export const App = () => {
