+++ src/components/DistanceForm.tsx
@@ -1,9 +1,5 @@
 import { useFormikContext } from "formik";
 
-import { FormState, RaceDistance, Unit } from "../types";
-import { getTotalTimeInSeconds, calculateDistance } from "../utils/calc";
-import { predefinedRaces, MILES_TO_KILOMETERS } from "../const";
-
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import {
@@ -13,6 +9,11 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select";
+
+import { predefinedRaces, MILES_TO_KILOMETERS } from "../const";
+import { FormState, RaceDistance, Unit } from "../types";
+import { getTotalTimeInSeconds, calculateDistance } from "../utils/calc";
+
 
 export const DistanceForm = () => {
   const { setFieldValue, values, handleChange } = useFormikContext<FormState>();
