+++ src/components/theme/theme-context.tsx
@@ -1,0 +1,16 @@
+import { createContext } from "react";
+
+import type { Theme } from "./types";
+
+type ThemeProviderState = {
+  theme: Theme;
+  setTheme: (theme: Theme) => void;
+};
+
+const initialState: ThemeProviderState = {
+  theme: "system",
+  setTheme: () => null,
+};
+
+export const ThemeProviderContext =
+  createContext<ThemeProviderState>(initialState);
