+++ src/components/theme/theme-toggle.tsx
@@ -2,7 +2,6 @@
 
 import { Moon, Sun } from "lucide-react";
 
-import { useTheme } from "@/hooks/use-theme";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
@@ -10,6 +9,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
+import { useTheme } from "@/hooks/use-theme";
 
 export function ThemeToggle() {
   const { setTheme } = useTheme();
