+++ src/components/theme/types.ts
@@ -1,0 +1,1 @@
+export type Theme = "light" | "dark" | "system";
