+++ src/utils/calc.test.ts
@@ -1,6 +1,7 @@
+import { calculateDistance, calculatePace, calculateTime } from "./calc";
+
 import { Unit } from "../types";
 
-import { calculateDistance, calculatePace, calculateTime } from "./calc";
 
 it("calculates pace", () => {
   expect(calculatePace(60 * 60, 1, Unit.MILES, Unit.MILES)).toStrictEqual({
