+++ src/utils/calc.ts
@@ -1,5 +1,5 @@
-import { Unit } from "../types";
 import { MILES_TO_KILOMETERS, KILOMETERS_TO_MILES } from "../const";
+import { Unit } from "../types";
 
 interface Time {
   seconds: number;
