+++ src/utils/generate-splits.test.ts
@@ -1,6 +1,7 @@
+import { generateSplits } from "./generate-splits";
+
 import { Unit } from "../types";
 
-import { generateSplits } from "./generate-splits";
 
 describe("generateSplits", () => {
   it("generates splits correctly for a given time and distance", () => {
