+++ vite.config.ts
@@ -1,9 +1,10 @@
 /// <reference types="vitest" />
 
+import react from "@vitejs/plugin-react-swc";
+import { defineConfig } from "vite";
+
 import path from "path";
 
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react-swc";
 
 export default defineConfig({
   plugins: [react()],
